pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ltg-net"
version = "0.1.0"
description = "Lagrangian Trajectory Guided Generative Weather Forecasting"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
  { name = "LTG-Net Research Team" }
]
dependencies = [
  "numpy>=1.19",
  "scipy>=1.5",
  "pandas>=1.1",
  "xarray>=0.20",
  "netCDF4>=1.5",
  "dask>=0.19",
  "torch>=1.13",
  "pyyaml>=3.13",
  "tqdm>=4.0",
  "matplotlib>=3.0",
  "scikit-learn>=1.0",
  "numba>=0.39",
  "cdsapi>=0.6",
  "h5py>=3.8",
  "requests>=2.31",
  "earthaccess>=0.11"
]

[project.scripts]
ltg-net = "ltg_net.cli:main"

[tool.setuptools.packages.find]
include = ["ltg_net*"]
